tests/wpt/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-single-value-in-body.html

<!doctype html>
<title>CSS Color Adjust Test: meta color-scheme in body should not apply</title>
<link rel="help" href="https://drafts.csswg.org/css-color-adjust-1/#color-scheme-meta">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/compute-root-color-scheme.js"></script>
<body>
  <meta name="color-scheme" content="dark">
</body>
<script>
  assert_root_color_scheme("normal", "Meta color-scheme in body does not apply.");
</script>
