third_party/blink/web_tests/external/wpt/css/css-color-adjust/meta/meta-color-scheme-single-value-in-body-expected.txt

This is a testharness.js-based test.
FAIL Meta color-scheme in body does not apply. assert_equals: expected (string) "normal" but got (undefined) undefined
Harness: the test ran to completion.
